+++ domain/currency.py
@@ -7,9 +7,3 @@ class CurrencyService:
         if currency == self._base:
             return amount
         return amount * self._rates[currency]
-
-
-# Пример использования:
-# currency_service = CurrencyService("KZT", {"EUR": 590, "USD": 500})
-# amount_in_usd = currency_service.convert(10000, "KZT")
-# print(f"10000 USD в KZT: {amount_in_usd}")